+++ sharding-proxy/src/test/java/io/shardingsphere/shardingproxy/config/AllConfigTests.java
@@ -22,6 +22,6 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
-@SuiteClasses(ProxyYamlConfigurationLoaderTest.class)
+@SuiteClasses(ShardingConfigurationLoaderTest.class)
 public final class AllConfigTests {
 }
